setting up authentication for api requests, check test and function
