+++ routes/products.js
@@ -1,0 +1,5 @@
+const express = require('express');
+const router = express.Router();
+
+
+module.exports = router
